python-backend/requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
bce-python-sdk==0.9.46
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
Deprecated==1.3.1
einops==0.8.1
filelock==3.20.0
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3
modelscope==1.31.0
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.1
paddlepaddle-gpu==3.0.0
paddlex==3.3.6
pandas==2.3.3
pdf2image==1.17.0
pikepdf==10.0.0
pillow==12.0.0
platformdirs==4.5.0
prettytable==3.16.0
protobuf==3.20.2
psutil==7.1.2
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
PyMuPDF==1.26.5
pypdfium2==5.0.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.10.23
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
wrapt==2.0.0
